rename class names
